Split up the annotation JSON presentation service
This nests the service in one level to make room for the formatters to join
it as collaborators.

This is a pretty straight split and doesn't try and tackle any of the weirdness
of these tests.
