+++ h/services/annotation_json_presentation/__init__.py
@@ -1,0 +1,6 @@
+from h.services.annotation_json_presentation.factory import (  # noqa: F401
+    annotation_json_presentation_service_factory,
+)
+from h.services.annotation_json_presentation.service import (  # noqa: F401
+    AnnotationJSONPresentationService,
+)
